requirements.txt
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.13.2
mysql-connector-python==8.0.19
mysqlclient==1.4.6
PyMySQL==0.9.3
python-dotenv==0.12.0
pytz==2019.1
redis==3.4.1
requests==2.22.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
